+++ .github/workflows/test.yml
@@ -5,14 +5,15 @@ on:
     branches: [main]
   push:
     branches: [main]
+
 jobs:
   run-node-tests:
     runs-on: ubuntu-latest
 
     strategy:
+      fail-fast: false
       matrix:
         node-version: [20.x, 22.x, 23.x]
-      fail-fast: false
 
     steps:
       - name: Checkout
@@ -27,8 +28,8 @@ jobs:
       - name: Install Node.js ${{ matrix.node-version }}
         uses: actions/setup-node@v4
         with:
-          node-version: ${{ matrix.node-version }}
           cache: "pnpm"
+          node-version: ${{ matrix.node-version }}
 
       - name: Install dependencies
         run: pnpm i
