+++ package.json
@@ -6,12 +6,19 @@
   "homepage": "https://github.com/kysely-org/kysely-typeorm",
   "author": "Igal Klebanov <igalklebanov@gmail.com>",
   "license": "MIT",
-  "main": "./dist/cjs/index.js",
-  "module": "./dist/esm/index.js",
+  "main": "./dist/index.js",
+  "module": "./dist/index.mjs",
+  "types": "./dist/index.d.ts",
   "exports": {
     ".": {
-      "import": "./dist/esm/index.js",
-      "require": "./dist/cjs/index.js"
+      "import": {
+        "types": "./dist/index.d.mts",
+        "default": "./dist/index.mjs"
+      },
+      "require": {
+        "types": "./dist/index.d.ts",
+        "default": "./dist/index.js"
+      }
     }
   },
   "files": [
@@ -30,27 +37,30 @@
     "dialect"
   ],
   "scripts": {
-    "biome": "biome",
     "build": "tsup",
-    "test": "vitest",
-    "type-check": "tsc --noEmit"
+    "check:exports": "attw . --pack",
+    "check:types": "tsc --noEmit",
+    "lint": "biome ci",
+    "prepublishOnly": "pnpm lint && pnpm build && pnpm check:exports",
+    "test": "vitest"
   },
   "peerDependencies": {
     "kysely": ">= 0.24.0 < 1",
     "typeorm": ">= 0.3.0 < 0.4.0"
   },
   "devDependencies": {
+    "@arethetypeswrong/cli": "^0.17.4",
     "@arktype/attest": "^0.46.0",
     "@biomejs/biome": "^1.9.4",
     "@tsconfig/node22": "^22.0.1",
     "@types/lodash": "^4.17.16",
-    "@types/node": "^22.13.10",
-    "@types/pg": "^8.11.12",
+    "@types/node": "^22.14.1",
+    "@types/pg": "^8.11.13",
     "better-sqlite3": "^11.9.1",
-    "kysely": "^0.27.5",
+    "kysely": "^0.28.1",
     "lodash": "^4.17.21",
-    "mssql": "^10.0.4",
-    "mysql2": "^3.12.0",
+    "mssql": "^11.0.1",
+    "mysql2": "^3.14.0",
     "pg": "^8.14.1",
     "reflect-metadata": "^0.2.2",
     "sqlite3": "^5.1.7",
