--- src/index.ts
@@ -1,4 +1,0 @@
-export * from './config.js'
-export * from './dialect.js'
-export * from './driver.js'
-export * from './kyselify.js'
