+++ src/kyselify.mts
@@ -1,4 +1,7 @@
-import type {Generated as KyselyGenerated, GeneratedAlways as KyselyGeneratedAlways} from 'kysely'
+import type {
+	Generated as KyselyGenerated,
+	GeneratedAlways as KyselyGeneratedAlways,
+} from 'kysely'
 
 /**
  * This is used to mark entity properties that have their values generated by TypeORM
@@ -56,7 +59,7 @@ import type {Generated as KyselyGenerated, GeneratedAlways as KyselyGeneratedAlw
  * ```
  */
 export type Generated<T> = T & {
-  readonly __kysely__generated__?: unique symbol
+	readonly __kysely__generated__?: unique symbol
 }
 
 /**
@@ -114,7 +117,7 @@ export type Generated<T> = T & {
  * ```
  */
 export type GeneratedAlways<T> = T & {
-  readonly __kysely__generated__always__?: unique symbol
+	readonly __kysely__generated__always__?: unique symbol
 }
 
 /**
@@ -178,7 +181,7 @@ export type GeneratedAlways<T> = T & {
  * ```
  */
 export type NonAttribute<T> = T & {
-  readonly __kysely__non__attribute__?: unique symbol
+	readonly __kysely__non__attribute__?: unique symbol
 }
 
 /**
@@ -284,17 +287,18 @@ export type NonAttribute<T> = T & {
  * ```
  */
 export type KyselifyEntity<E> = {
-  [K in keyof E as E[K] extends (...args: any) => any
-    ? never
-    : '__kysely__non__attribute__' extends keyof E[K]
-      ? never
-      : K]-?: '__kysely__generated__' extends keyof E[K]
-    ? E[K] extends Generated<infer T>
-      ? KyselyGenerated<Exclude<T, undefined>>
-      : never
-    : '__kysely__generated__always__' extends keyof E[K]
-      ? E[K] extends GeneratedAlways<infer T>
-        ? KyselyGeneratedAlways<Exclude<T, undefined>>
-        : never
-      : Exclude<E[K], undefined>
+	// biome-ignore lint/suspicious/noExplicitAny: it's fine here.
+	[K in keyof E as E[K] extends (...args: any) => any
+		? never
+		: '__kysely__non__attribute__' extends keyof E[K]
+			? never
+			: K]-?: '__kysely__generated__' extends keyof E[K]
+		? E[K] extends Generated<infer T>
+			? KyselyGenerated<Exclude<T, undefined>>
+			: never
+		: '__kysely__generated__always__' extends keyof E[K]
+			? E[K] extends GeneratedAlways<infer T>
+				? KyselyGeneratedAlways<Exclude<T, undefined>>
+				: never
+			: Exclude<E[K], undefined>
 }
