--- src/type-utils.ts
@@ -1,3 +1,0 @@
-export function isObject(thing: unknown): thing is Record<string, any> {
-  return typeof thing === 'object' && thing !== null && !Array.isArray(thing)
-}
