+++ tsup.config.mts
@@ -1,0 +1,9 @@
+import { defineConfig } from 'tsup'
+
+export default defineConfig({
+	clean: true,
+	dts: true,
+	entry: ['./src/index.mts'],
+	format: ['cjs', 'esm'],
+	shims: true,
+})
