--- tsup.config.ts
@@ -1,11 +1,0 @@
-import {defineConfig} from 'tsup'
-
-export default defineConfig({
-  clean: true,
-  dts: true,
-  entry: ['./src/index.ts'],
-  format: ['cjs', 'esm'],
-  legacyOutput: true,
-  outDir: 'dist',
-  sourcemap: true,
-})
